+++ src/main/java/io/lettuce/core/RedisAsyncCommandsImpl.java
@@ -4,6 +4,8 @@ import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.async.RedisAsyncCommands;
 import io.lettuce.core.cluster.api.async.RedisClusterAsyncCommands;
 import io.lettuce.core.codec.RedisCodec;
+import io.lettuce.core.json.JsonParser;
+import io.lettuce.core.json.JsonParserRegistry;
 
 /**
  * An asynchronous and thread-safe API for a Redis connection.
@@ -15,6 +17,8 @@ import io.lettuce.core.codec.RedisCodec;
 public class RedisAsyncCommandsImpl<K, V> extends AbstractRedisAsyncCommands<K, V>
         implements RedisAsyncCommands<K, V>, RedisClusterAsyncCommands<K, V> {
 
+    private final RedisCodec<K,V> codec;
+
     /**
      * Initialize a new instance.
      *
@@ -24,11 +28,22 @@ public class RedisAsyncCommandsImpl<K, V> extends AbstractRedisAsyncCommands<K, 
      */
     public RedisAsyncCommandsImpl(StatefulRedisConnection<K, V> connection, RedisCodec<K, V> codec) {
         super(connection, codec);
+        this.codec = codec;
     }
 
     @Override
     public StatefulRedisConnection<K, V> getStatefulConnection() {
         return (StatefulRedisConnection<K, V>) super.getConnection();
+    }
+
+    @Override
+    public JsonParser<K, V> getJsonParser() {
+        return JsonParserRegistry.getJsonParser(this.codec);
+    }
+
+    @Override
+    public void setJsonParser(JsonParser<K, V> jsonParser) {
+        throw new UnsupportedOperationException("Setting a custom JsonParser is not supported");
     }
 
 }
