+++ src/main/java/io/lettuce/core/RedisReactiveCommandsImpl.java
@@ -4,6 +4,8 @@ import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.reactive.RedisReactiveCommands;
 import io.lettuce.core.cluster.api.reactive.RedisClusterReactiveCommands;
 import io.lettuce.core.codec.RedisCodec;
+import io.lettuce.core.json.JsonParser;
+import io.lettuce.core.json.JsonParserRegistry;
 
 /**
  * A reactive and thread-safe API for a Redis Sentinel connection.
@@ -15,6 +17,8 @@ import io.lettuce.core.codec.RedisCodec;
 public class RedisReactiveCommandsImpl<K, V> extends AbstractRedisReactiveCommands<K, V>
         implements RedisReactiveCommands<K, V>, RedisClusterReactiveCommands<K, V> {
 
+    private final RedisCodec<K,V> codec;
+
     /**
      * Initialize a new instance.
      *
@@ -24,11 +28,22 @@ public class RedisReactiveCommandsImpl<K, V> extends AbstractRedisReactiveComman
      */
     public RedisReactiveCommandsImpl(StatefulRedisConnection<K, V> connection, RedisCodec<K, V> codec) {
         super(connection, codec);
+        this.codec = codec;
     }
 
     @Override
     public StatefulRedisConnection<K, V> getStatefulConnection() {
         return (StatefulRedisConnection<K, V>) super.getConnection();
+    }
+
+    @Override
+    public JsonParser<K, V> getJsonParser() {
+        return JsonParserRegistry.getJsonParser(this.codec);
+    }
+
+    @Override
+    public void setJsonParser(JsonParser<K, V> jsonParser) {
+        throw new UnsupportedOperationException("Setting a custom JsonParser is not supported");
     }
 
 }
