+++ src/main/java/io/lettuce/core/api/async/RedisAsyncCommands.java
@@ -22,6 +22,7 @@ package io.lettuce.core.api.async;
 import io.lettuce.core.RedisFuture;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.cluster.api.async.RedisClusterAsyncCommands;
+import io.lettuce.core.json.JsonParser;
 
 /**
  * A complete asynchronous and thread-safe Redis API with 400+ Methods.
@@ -80,5 +81,9 @@ public interface RedisAsyncCommands<K, V> extends BaseRedisAsyncCommands<K, V>, 
      */
     @Deprecated
     StatefulRedisConnection<K, V> getStatefulConnection();
+
+    JsonParser<K, V> getJsonParser();
+
+    void setJsonParser(JsonParser<K, V> jsonParser);
 
 }
