+++ src/main/java/io/lettuce/core/api/reactive/RedisReactiveCommands.java
@@ -21,6 +21,7 @@ package io.lettuce.core.api.reactive;
 
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.cluster.api.reactive.RedisClusterReactiveCommands;
+import io.lettuce.core.json.JsonParser;
 import reactor.core.publisher.Mono;
 
 /**
@@ -80,5 +81,9 @@ public interface RedisReactiveCommands<K, V> extends BaseRedisReactiveCommands<K
      */
     @Deprecated
     StatefulRedisConnection<K, V> getStatefulConnection();
+
+    JsonParser<K, V> getJsonParser();
+
+    void setJsonParser(JsonParser<K, V> jsonParser);
 
 }
